Implemented changes to simple search box (#13)
* Implemented changes to simple search box

Implemented wild-card search (with `?` and `*`)
Removed existing options from simple search. Will now search on the precise string (possibly with wildcards).
Added the option to choose the collection to search ("Formulae" or "Charters").
Now raises ValidationErrors when a multiword search term is given with a wildcard or when neither collection check box is selected.
Added tests for simple search form and the function.