First changes after the evaluation
A click on a raised footnote in the text will now go to the footnote on the right side of the reading screen. It will also quickly change the background color to yellow to draw attention to it. - I took away the X to close the notes since there is no easy way to open them back up when they are closed. - Also finished with the French translations.